requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-cors-headers==3.10.1
djangorestframework==3.13.1
ecdsa==0.17.0
idna==3.3
opentok==3.1.0
pyasn1==0.4.8
python-jose==3.3.0
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
